feat: create roaster endpoint, add node api port env var, and express json middleware into index.ts
