+++ backend/.env.example
@@ -6,3 +6,5 @@ TWITTER_ACCESS_TOKEN_SECRET=
 OPENAI_API_KEY=
 
 GITHUB_TOKEN=
+
+NODE_API_PORT=
